backend/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flake8==6.0.0
flake8-annotations==3.0.1
flake8-commas==2.1.0
flake8-quotes==3.3.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
packaging==23.1
Pillow==8.1.0
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.26.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
toml==0.10.2
urllib3==1.26.15